大喜利関連のドメインクラスを集約
