+++ spring_prj/src/main/java/com/c4ccup/laugh/controller/OogiriAnswerResponse.java
@@ -18,92 +18,110 @@ public class OogiriAnswerResponse {
     private String answerContent;
     /** 回答作成日時 */
     private LocalDateTime answerCreatedAt;
+    /** 回答削除日時 */
+    private LocalDateTime answerDeletedAt;
 
     /** リアクションリスト */
     private List<OogiriReactionResponse> reactions;
 
     /**
-     * 回答IDを取得
-     * 
-     * @return answerId
+     * 回答IDを取得します。
+     * @return 回答ID
      */
     public int getAnswerId() {
         return answerId;
     }
 
     /**
-     * @param answerId セットする 回答ID
+     * 回答IDを設定します。
+     * @param answerId 回答ID
      */
     public void setAnswerId(int answerId) {
         this.answerId = answerId;
     }
 
     /**
-     * 回答ユーザーIDを取得
-     * 
-     * @return answerUserId
+     * 回答ユーザーIDを取得します。
+     * @return 回答ユーザーID
      */
     public int getAnswerUserId() {
         return answerUserId;
     }
 
     /**
-     * @param answerUserId セットする 回答ユーザーID
+     * 回答ユーザーIDを設定します。
+     * @param answerUserId 回答ユーザーID
      */
     public void setAnswerUserId(int answerUserId) {
         this.answerUserId = answerUserId;
     }
 
     /**
-     * 回答内容を取得
-     * 
-     * @return answerContent
+     * 回答内容を取得します。
+     * @return 回答内容
      */
     public String getAnswerContent() {
         return answerContent;
     }
 
     /**
-     * @param answerContent セットする 回答内容
+     * 回答内容を設定します。
+     * @param answerContent 回答内容
      */
     public void setAnswerContent(String answerContent) {
         this.answerContent = answerContent;
     }
 
     /**
-     * 回答作成日時を取得
-     * 
-     * @return answerCreatedAt
+     * 回答作成日時を取得します。
+     * @return 回答作成日時
      */
     public LocalDateTime getAnswerCreatedAt() {
         return answerCreatedAt;
     }
 
     /**
-     * @param answerCreatedAt セットする 回答作成日時
+     * 回答作成日時を設定します。
+     * @param answerCreatedAt 回答作成日時
      */
     public void setAnswerCreatedAt(LocalDateTime answerCreatedAt) {
         this.answerCreatedAt = answerCreatedAt;
     }
 
     /**
-     * リアクションリストを取得
-     * 
-     * @return reactions
+     * 回答削除日時を取得します。
+     * @return 回答削除日時
+     */
+    public LocalDateTime getAnswerDeletedAt() {
+        return answerDeletedAt;
+    }
+
+    /**
+     * 回答削除日時を設定します。
+     * @param answerDeletedAt 回答削除日時
+     */
+    public void setAnswerDeletedAt(LocalDateTime answerDeletedAt) {
+        this.answerDeletedAt = answerDeletedAt;
+    }
+
+    public OogiriAnswerResponse() {
+        this.reactions = new ArrayList<>();
+    }
+
+    /**
+     * リアクションリストを取得します。
+     * @return リアクションリスト
      */
     public List<OogiriReactionResponse> getReactions() {
         return reactions;
     }
 
     /**
-     * @param reactions セットする リアクションリスト
+     * リアクションリストを設定します。
+     * @param reactions リアクションリスト
      */
     public void setReactions(List<OogiriReactionResponse> reactions) {
         this.reactions = reactions;
-    }
-
-    public OogiriAnswerResponse() {
-        this.reactions = new ArrayList<>();
     }
 
 }
